please read the comments
articles for CI/CD: first two provided links are mostly similar (regarding the second link itself, it is just a clone of redHat article) => it would be nice if you remove one of them.
for addressing folders: imply that addressing is done from the root. not from where the .yml files are located. also please make the word "folder" BOLD.  
also it seems that some "tips" are not formatted properly, so that they are not shown as they should.